--- scripted.sbt
@@ -1,7 +1,0 @@
-ScriptedPlugin.scriptedSettings
-
-scriptedLaunchOpts := { scriptedLaunchOpts.value ++
-  Seq("-Xmx1024M", "-Dplugin.version=" + version.value)
-}
-
-scriptedBufferLog := false
