--- src/main/scala/com/chatwork/sbt/aws/AwsKeys.scala
@@ -1,9 +1,0 @@
-package com.chatwork.sbt.aws
-
-import com.amazonaws.services.elasticbeanstalk.model.{ ConfigurationOptionSetting, OptionSpecification }
-
-case class EbConfigurationTemplate(name: String, description: String,
-                                   solutionStackName: String,
-                                   optionSettings: Seq[ConfigurationOptionSetting],
-                                   optionsToRemoves: Seq[OptionSpecification],
-                                   recreate: Boolean)
