--- src/sbt-test/sbt-aws-plugin/cloud-formation/aws/cfn/templates/S3.template
@@ -1,28 +1,0 @@
-{
-     "AWSTemplateFormatVersion" : "2010-09-09",
-     "Parameters": {
-        "S3BucketName0": {
-            "Description": "Bucket Name on S3",
-            "Type" : "String"
-        },
-        "S3BucketName1": {
-            "Description": "Bucket Name on S3",
-            "Type" : "String"
-        }
-     },
-     "Resources" : {
-       "S3Bucket0" : {
-         "Type" : "AWS::S3::Bucket",
-         "Properties" : {
-            "BucketName" : { "Ref" : "S3BucketName0" }
-         }
-       },
-       "S3Bucket1" : {
-         "Type" : "AWS::S3::Bucket",
-         "Properties" : {
-            "BucketName" : { "Ref" : "S3BucketName1" }
-         }
-       }
-     },
-     "Description": "test"
-}
