--- src/sbt-test/sbt-aws-plugin/cloud-formation/test
@@ -1,2 +1,0 @@
-> aws::cfnStackCreateOrUpdateAndWait
-
