--- src/sbt-test/sbt-aws-plugin/simple/project/plugins.sbt
@@ -1,15 +1,0 @@
-logLevel := Level.Info
-
-resolvers ++= Seq(
-  "Sonatype OSS Snapshot Repository" at "https://oss.sonatype.org/content/repositories/snapshots/",
-  "Sonatype OSS Release Repository" at "https://oss.sonatype.org/content/repositories/releases/",
-  "Typesafe Releases" at "http://repo.typesafe.com/typesafe/releases/"
-)
-
-{
-  val pluginVersion = System.getProperty("plugin.version")
-  if(pluginVersion == null)
-    throw new RuntimeException("""|The system property 'plugin.version' is not defined.
-                                 |Specify this property using the scriptedLaunchOpts -D.""".stripMargin)
-  else addSbtPlugin("com.chatwork" % "sbt-aws" % pluginVersion)
-}
