--- src/sbt-test/sbt-aws-plugin/simple/test
@@ -1,1 +1,0 @@
-> aws::s3Upload
